map

This function produce different kinds of maps for the rmap package. Each figure is accompanied with a csv table.

map(
  data = NULL,
  fillColumn = NULL,
  shapeColumn = NULL,
  shape = NULL,
  fileName = "map",
  save = T,
  theme = NULL,
  show = T,
  folder = paste(getwd(), "/outputs", sep = ""),
  labels = F,
  labelRepel = 0,
  labelColor = "black",
  labelSize = 2,
  labelAlpha = 1,
  labelFill = NA,
  labelBorderSize = NA,
  shapeFolder = NULL,
  shapeFile = NULL,
  subRegCol = "subRegion",
  nameAppend = "",
  legendTitle = NULL,
  legendType = "kmeans",
  legendBreaksn = 5,
  legendFixedBreaks = NULL,
  animate = T,
  fps = 1,
  crop = T,
  crop_to_underLayer = F,
  crop_to_overLayer = F,
  underLayer = NULL,
  underLayerColor = "gray40",
  underLayerFill = "gray90",
  underLayerLwd = 0.1,
  underLayerAlpha = 1,
  underLayerLabels = F,
  overLayerLabels = F,
  overLayer = NULL,
  overLayerColor = "gray40",
  overLayerFill = NA,
  overLayerLwd = 0.2,
  overLayerAlpha = 0,
  zoom = 0,
  zoomx = NULL,
  zoomy = NULL,
  asp = 1.2,
  projX = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  width = 7,
  height = 8,
  scenRef = NULL,
  scenDiff = NULL,
  scaleRange = NULL,
  scaleRangeDiffAbs = NULL,
  scaleRangeDiffPrcnt = NULL,
  xRef = NULL,
  xDiff = NULL,
  scaleRangexDiffAbs = NULL,
  scaleRangexDiffPrcnt = NULL,
  col = NULL,
  row = NULL,
  title = NULL,
  numeric2Cat_list = NULL,
  pdfpng = "png",
  legendDigitsOverride = NULL,
  legendSingleColor = "white",
  legendSingleValue = F,
  palette = NULL,
  paletteDiff = "pal_div_BluRd",
  colorNA = "gray50",
  showNA = F,
  ncol = 3,
  size = 16,
  alpha = 1,
  background = F,
  transparent = T,
  legendShow = T,
  diffOnly = F,
  forceFacets = F
)

Arguments

data
Default = NULL,

fillColumn
(Optional). Default = NULL. Only for direct map plotting.

shapeColumn
(Optional). Default = NULL. If different from subRegion.

shape
Default = NULL, Cusotm shape can be provided as a SpatialPolygonDataFrame with features corresponding to subRegion columns in the data provided.

fileName
(Optional). Default = "map". Only for direct map plotting.

save
(Optional). Default = T. Only for direct map plotting.

theme
Default = NULL,

show
Default = T. Print maps in console as they are processed.

folder
Default = paste(getwd(),"/outputs",sep=""),

labels
Default = F,

labelRepel
Default = 0,

labelColor
Default = "black",

labelSize
Default = 3

labelAlpha
Default = 0.7,

labelFill
Default = NA,

labelBorderSize
Default = NA

shapeFolder
Default = paste(getwd(),"/dataFiles/gis/admin_gadm36",sep=""),

shapeFile
Default = paste("gadm36_1",sep=""),

subRegCol
Default ="subRegion",

nameAppend
Default =""

legendTitle
Default = NULL

legendType
Default ="kmeans", Options include c("pretty","kmeans","freescale","all")

legendBreaksn
Default = "5",

legendFixedBreaks
Default = NULL,

animate
Default = T,

fps
Default = 1,

crop
Default = T. This crops the map to the extent of your data regions. If false will zoom out to the extent of the larget layer.

crop_to_underLayer
Default = F. Crop to the underLayer boundary provided.

crop_to_overLayer
Default = F. Crop to the overLayer boundary provided.

underLayer
Default = NULL

underLayerColor
Default = "gray40"

underLayerFill
Default = "gray90"

underLayerLwd
Default = 0.5

underLayerAlpha
Default = 1

underLayerLabels
Default = F

overLayerLabels
Default = F

overLayer
Default = NULL

overLayerColor
Default = "gray40"

overLayerFill
Default = NA

overLayerLwd
Default = 0.5

overLayerAlpha
Default = 0

zoom
Default =-1. Zoom into or out of map. Positive values zoom in and negative out.

zoomx
Default = NULL. Zoom into or out of map along x. Positive values zoom in and negative out.

zoomy
Default = NULL. Zoom into or out of map along y. Positive values zoom in and negative out.

asp
Default = 1.2. Aspect ratio of lat and lon.

projX
Default = projX="+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

width
Default =9

height
Default =7

scenRef
Reference Scenario. Default = NULL

scenDiff
Scenarios to Diff. Default = NULL

scaleRange
Default NULL. A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min

scaleRangeDiffAbs
Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min

scaleRangeDiffPrcnt
Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min

xRef
Reference year. Default = NULL

xDiff
years to Diff. Default = NULL

scaleRangexDiffAbs
Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min

scaleRangexDiffPrcnt
Default =NULL, A vector with c(max,min) (Applied to all params) or a dataframe with cols param, max, min

col
Default ="multiFacetRow",

row
Default ="multiFacetCol",

title
Default=NULL

numeric2Cat_list
Default=NULL,

pdfpng
Save IO figures as pdf or png. Type=String. Options: 'pdf' or 'png'. Default = 'png'

legendDigitsOverride
Default=NULL

legendSingleColor
Default="white"

legendSingleValue
Default=F. Change to True to get default single value or provide a numeric value.

palette
Default = NULL

paletteDiff
Default = "pal_div_BrGn"

colorNA
Default = "gray50"

showNA
Default = T

ncol
Default = 3. Number of columns to wrap maps

size
Default = 12. Text size of plots.

alpha
Default = 1. Transparency of fill colors.

background
Default = F. Add background water color, border and default underlayer map.

transparent
Default = T. To make map background transparent for maps without backgrounds.

legendShow
Default = T

diffOnly
Default = F. Only run diff plots and not individual scenarios.

forceFacets
Default = F. Used to force facet label for single scenario which is usually dropped.

Value

Returns the formatted data used to produce chart

A list of maps