US 52 States Compact Includes Alaska, Hawaii and Puerto Rico as well as DC (Re-positioned)

US 52 States Compact Includes Alaska, Hawaii and Puerto Rico as well as DC (Re-positioned)

Usage,

mapUS52Compactdf

Format

A SpatialPolygonsDataFrame

Source

US Census bureau. https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html

Examples

if (FALSE) {
 library(sp); library(rmap); library(ggplot2)
 ggplot() +
 geom_polygon(data = mapUS52Compactdf, aes(x = long, y = lat, group = group),
             colour = "black", fill = NA)
 head(mapUS52Compactdf)
}