map_plot

This function produce different kinds of maps for the rmap package. Each figure is accompanied with a csv table.

map_plot(
  data = NULL,
  fillColumn = NULL,
  theme = NULL,
  fillPalette = "Spectral",
  labels = F,
  labelForce = 0,
  labelColor = "black",
  labelSize = 3,
  labelAlpha = 1,
  labelFill = NA,
  labelBorderSize = NA,
  width = 9,
  height = 7,
  legendShow = F,
  legendTitle = NULL,
  legendBreaksn = 5,
  legendBreaks = NULL,
  pdfpng = "png",
  underLayer = NULL,
  underLayerColor = "gray40",
  underLayerFill = "gray90",
  underLayerLwd = 0.1,
  underLayerAlpha = 1,
  underLayerLabels = F,
  overLayerLabels = F,
  overLayer = NULL,
  overLayerColor = "gray40",
  overLayerFill = NA,
  overLayerLwd = 0.2,
  overLayerAlpha = 0,
  zoom = -1,
  zoomx = NULL,
  zoomy = NULL,
  asp = 1.2,
  save = T,
  fileName = "map",
  folder = paste(getwd(), "/outputs", sep = ""),
  facetRows = NULL,
  facetCols = NULL,
  mapTitle = NULL,
  numeric2Cat_list = NULL,
  legendDigits = NULL,
  legendDigitsOverride = NULL,
  legendSingleColor = "white",
  legendSingleValue = NULL,
  colorNA = "gray50",
  showNA = F,
  ncol = 3,
  size = 12,
  alpha = 1,
  background = F,
  cropToBoundary = T,
  transparent = F
)

Arguments

data
Default = NULL

fillColumn
Default = NULL # Or give column with data

theme
Default = NULL

fillPalette
Default = "Spectral"

labels
Default = FALSE

labelForce
Default = 0,

labelColor
Default = "black",

labelSize
Default = 3

labelAlpha
Default = 0.7,

labelFill
Default = NA,

labelBorderSize
Default = NA

width
Default = 9

height
Default = 7

legendShow
Default = F

legendTitle
Default = NULL

legendBreaksn
Default = "5"

legendBreaks
Default = NULL

pdfpng
Default = "png"

underLayer
Default = NULL

underLayerColor
Default = "gray40"

underLayerFill
Default = "gray40"

underLayerLwd
Default = 0.5

underLayerAlpha
Default = 1

underLayerLabels
Default = F

overLayerLabels
Default = F

overLayer
Default = NULL

overLayerColor
Default = "gray40"

overLayerFill
Default = NA

overLayerLwd
Default = 0.5

overLayerAlpha
Default = 0

zoom
Default =-1. Zoom into or out of map. Positive values zoom in and negative out.

zoomx
Default = NULL. Zoom into or out of map along x. Positive values zoom in and negative out.

zoomy
Default = NULL. Zoom into or out of map along y. Positive values zoom in and negative out.

asp
Default = 1.2. Aspect ratio of lat and lon.

save
Default = T

fileName
Default = "map"

folder
Default = paste(getwd(),"/outputs",sep Default = "")

facetRows
Default=NULL

facetCols
Default=NULL

mapTitle
Default=NULL

numeric2Cat_list
Default=NULL

legendDigits
Default = NULL

legendDigitsOverride
Default=NULL

legendSingleColor
Default="white"

legendSingleValue
Default=NULL

colorNA
Default = "gray50"

showNA
Default = T

ncol
Default = 3. Number of columns to wrap maps

size
Default = 12. Text size of plots.

alpha
Default = 1. Transparency of fill colors.

background
Default = F. Add background water color, border and default underlayer map.

cropToBoundary
Default = T. Crop to boundary data.

transparent
Default = F. To make map background transparent for maps without backgrounds.

Value

Returns the formatted data used to produce chart